+++ NEWS.md
@@ -8,6 +8,10 @@ NEW FEATURES :
 * `createDistrict()` uses a specific endpoint to create the district
 * `editDistrict()` allows the user to edit a district. Use of a specific endpoint in API mode.
 
+ENHANCEMENT :
+
+* `createBindingConstraint()`/`editBindingCOnstraint()` : remove depency to `readClusterDesc()` from antaresRead package
+
 BUGFIXES :    
 
 * `.get_version_solver_from_path_solver()` get solver version by running command line instead of searching in the path.